fix add zap scripts
