+++ scripts/setVaultUrls.ts
@@ -6,7 +6,7 @@ import { sortVaultKeys } from './common/vault-fields.ts';
 import type { VaultConfig } from '../src/features/data/apis/config-types.ts';
 import { getTokenById } from './common/tokens.ts';
 import { type TokenEntity } from '../src/features/data/entities/token.ts';
-import { type CurveApiPoolWithMetadata, getCurvePools } from './addCurveZap.ts';
+import { type CurveApiPoolWithMetadata, getCurvePools } from './zaps/curve.ts';
 import { mapValuesAsync } from './common/utils.ts';
 import { type ChainEntity } from '../src/features/data/entities/chain.ts';
 import { entries } from '../src/helpers/object.ts';
